Fix build scriptlet append/prepend interaction with Buildsystem
The append and prepend modes got added before the declarative
Buildsystem, and did not get thoroughly tested with it. The existing
%build -a test didn't actually work but automake handling the build
in %install masked the issue embarrasingly. As the Buildsystem is parsed
after everything else, there's no way the previous append/prepend
implementation could work correctly with it.

Do what we should've done from the start: collect any prepend/append
stuff into separate data structures and apply them after everything
else has been parsed. This also lifts the artificial sounding
restriction wrt the corresponding main section:it's really the right
thing to do, even if it's a bit more code.

Make the tests wrt buildsystem much more thorough to ensure it all
really works, blush.

Fixes: #3024
(cherry picked from commit bf46dcf3646e1c296c1a5a473aece07cff6682bf)

+++ tests/data/SPECS/amhello.spec
@@ -21,13 +21,35 @@ Release: 1
 %description
 %{summary}
 
+%build -p
+touch pre1
+
+%build -p
+touch pre2
+
 %build -a
+test -f pre1 || exit 1
+test -f pre2 || exit 1
 cat << EOF > README.distro
 Add some distro specific notes.
 EOF
 
+%install -p
+mkdir -p ${RPM_BUILD_ROOT}/%{_sysconfdir}/hello.p
+
+%install -a
+test -d ${RPM_BUILD_ROOT}/%{_sysconfdir}/hello.p || exit 1
+mkdir -p ${RPM_BUILD_ROOT}/%{_sysconfdir}/hello.d
+
+%install -a
+test -d ${RPM_BUILD_ROOT}/%{_sysconfdir}/hello.d || exit 1
+mkdir -p ${RPM_BUILD_ROOT}/%{_sysconfdir}/hello.a
+
 %files
 %doc README.distro
+%{_sysconfdir}/hello.a/
+%{_sysconfdir}/hello.d/
+%{_sysconfdir}/hello.p/
 %if %{with alt}
 /alt/%{_bindir}/alt-hello
 /alt/%{_docdir}/%{name}
